+++ packages/widgetbook/lib/src/component/tooltip_use_case.dart
@@ -20,8 +20,8 @@ Widget buildWdsTooltipUseCase(BuildContext context) {
 }
 
 Widget _buildPlaygroundSection(BuildContext context) {
-  final message = context.knobs.string(
-    label: 'message',
+  final label = context.knobs.string(
+    label: 'label',
     description: '툴팁에 표시될 메시지를 입력해 주세요',
     initialValue: '툴팁 메시지입니다.',
   );
@@ -46,7 +46,7 @@ Widget _buildPlaygroundSection(BuildContext context) {
 
   return WidgetbookPlayground(
     info: [
-      'message: "$message"',
+      'label: "$label"',
       'hasArrow: $hasArrow',
       'hasCloseButton: $hasCloseButton',
       'alignment: ${alignment.name}',
@@ -65,7 +65,7 @@ Widget _buildPlaygroundSection(BuildContext context) {
       height: 200,
       child: Center(
         child: WdsTooltip(
-          message: message,
+          label: Text(label),
           hasArrow: hasArrow,
           hasCloseButton: hasCloseButton,
           alignment: alignment,
@@ -87,9 +87,11 @@ Widget _buildDemonstrationSection(BuildContext context) {
           mainAxisSize: MainAxisSize.min,
           spacing: 24,
           children: [
-            WdsTooltip(message: '화살표가 있는 툴팁'),
             WdsTooltip(
-              message: '화살표가 없는 툴팁',
+              label: Text('화살표가 있는 툴팁'),
+            ),
+            WdsTooltip(
+              label: Text('화살표가 없는 툴팁'),
               hasArrow: false,
             ),
           ],
@@ -98,25 +100,71 @@ Widget _buildDemonstrationSection(BuildContext context) {
       const SizedBox(height: 32),
       WidgetbookSubsection(
         title: 'hasCloseButton',
-        labels: ['true', 'false'],
+        labels: const ['true', 'false'],
         content: Wrap(
           spacing: 16,
           runSpacing: 16,
           children: [
             WdsTooltip(
-              message: '화살표가 있고 닫기 버튼이 있는 툴팁',
+              label: const Text.rich(
+                TextSpan(
+                  children: [
+                    TextSpan(text: '화살표가 있고 '),
+                    TextSpan(
+                      text: '닫기 버튼이',
+                      style: TextStyle(color: WdsColors.primary),
+                    ),
+                    TextSpan(text: ' 있는 툴팁'),
+                  ],
+                ),
+              ),
               hasCloseButton: true,
               onClose: () {},
             ),
-            const WdsTooltip(message: '화살표가 있고 닫기 버튼이 없는 툴팁'),
+            const WdsTooltip(
+              label: Text.rich(
+                TextSpan(
+                  children: [
+                    TextSpan(text: '화살표가 있고 '),
+                    TextSpan(
+                      text: '닫기 버튼이',
+                      style: TextStyle(color: WdsColors.primary),
+                    ),
+                    TextSpan(text: ' 있는 툴팁'),
+                  ],
+                ),
+              ),
+            ),
             WdsTooltip(
-              message: '화살표가 없고 닫기 버튼이 있는 툴팁',
+              label: const Text.rich(
+                TextSpan(
+                  children: [
+                    TextSpan(text: '화살표가 있고 '),
+                    TextSpan(
+                      text: '닫기 버튼이',
+                      style: TextStyle(color: WdsColors.primary),
+                    ),
+                    TextSpan(text: ' 있는 툴팁'),
+                  ],
+                ),
+              ),
               hasCloseButton: true,
               onClose: () {},
               hasArrow: false,
             ),
             const WdsTooltip(
-              message: '화살표가 없고 닫기 버튼이 없는 툴팁',
+              label: Text.rich(
+                TextSpan(
+                  children: [
+                    TextSpan(text: '화살표가 있고 '),
+                    TextSpan(
+                      text: '닫기 버튼이',
+                      style: TextStyle(color: WdsColors.primary),
+                    ),
+                    TextSpan(text: ' 있는 툴팁'),
+                  ],
+                ),
+              ),
               hasArrow: false,
             ),
           ],
@@ -171,12 +219,12 @@ Widget _buildDemonstrationSection(BuildContext context) {
                           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                           children: [
                             WdsTooltip(
-                              message: 'topLeft',
+                              label: Text('topLeft'),
                               alignment: WdsTooltipAlignment.topLeft,
                             ),
-                            WdsTooltip(message: 'topCenter'),
+                            WdsTooltip(label: Text('topCenter')),
                             WdsTooltip(
-                              message: 'topRight',
+                              label: Text('topRight'),
                               alignment: WdsTooltipAlignment.topRight,
                             ),
                           ],
@@ -190,15 +238,15 @@ Widget _buildDemonstrationSection(BuildContext context) {
                               spacing: 16,
                               children: [
                                 WdsTooltip(
-                                  message: 'leftTop',
+                                  label: Text('leftTop'),
                                   alignment: WdsTooltipAlignment.leftTop,
                                 ),
                                 WdsTooltip(
-                                  message: 'leftCenter',
+                                  label: Text('leftCenter'),
                                   alignment: WdsTooltipAlignment.leftCenter,
                                 ),
                                 WdsTooltip(
-                                  message: 'leftBottom',
+                                  label: Text('leftBottom'),
                                   alignment: WdsTooltipAlignment.leftBottom,
                                 ),
                               ],
@@ -208,15 +256,15 @@ Widget _buildDemonstrationSection(BuildContext context) {
                               spacing: 16,
                               children: [
                                 WdsTooltip(
-                                  message: 'rightTop',
+                                  label: Text('rightTop'),
                                   alignment: WdsTooltipAlignment.rightTop,
                                 ),
                                 WdsTooltip(
-                                  message: 'rightCenter',
+                                  label: Text('rightCenter'),
                                   alignment: WdsTooltipAlignment.rightCenter,
                                 ),
                                 WdsTooltip(
-                                  message: 'rightBottom',
+                                  label: Text('rightBottom'),
                                   alignment: WdsTooltipAlignment.rightBottom,
                                 ),
                               ],
@@ -228,15 +276,15 @@ Widget _buildDemonstrationSection(BuildContext context) {
                           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                           children: [
                             WdsTooltip(
-                              message: 'bottomLeft',
+                              label: Text('bottomLeft'),
                               alignment: WdsTooltipAlignment.bottomLeft,
                             ),
                             WdsTooltip(
-                              message: 'bottomCenter',
+                              label: Text('bottomCenter'),
                               alignment: WdsTooltipAlignment.bottomCenter,
                             ),
                             WdsTooltip(
-                              message: 'bottomRight',
+                              label: Text('bottomRight'),
                               alignment: WdsTooltipAlignment.bottomRight,
                             ),
                           ],